my-first-app/requirements.txt
psycopg2-binary
flask
# Flask-SQLAlchemy==2.5.1
# flask-migrate